requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.3
gunicorn==21.2.0
lxml==5.1.0
packaging==24.0
psycopg2==2.9.9
python-docx==1.1.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
whitenoise==6.6.0
